Add handling of bad values for AdvConfig

+++ core/src/test/java/com/pipai/adv/AdvConfigTest.java
@@ -12,13 +12,16 @@ import com.pipai.test.libgdx.GdxTestUtilsKt;
 public class AdvConfigTest extends GdxMockedTest {
 
     public static final FileHandle TEST_FILE = GdxTestUtilsKt.getTestResourceFileHandle(AdvConfigTest.class, "config.properties");
+    public static final FileHandle BAD_TEST_FILE = GdxTestUtilsKt.getTestResourceFileHandle(AdvConfigTest.class, "bad-config.properties");
     public static final FileHandle TEST_WRITE_FILE = Gdx.files.local("test-config.properties");
 
     @After
     public void teardown() {
-        boolean deleted = TEST_WRITE_FILE.delete();
-        if (!deleted) {
-            throw new IllegalStateException("Test file was not deleted");
+        if (TEST_WRITE_FILE.exists()) {
+            boolean deleted = TEST_WRITE_FILE.delete();
+            if (!deleted) {
+                throw new IllegalStateException("Test file was not deleted");
+            }
         }
     }
 
@@ -31,6 +34,12 @@ public class AdvConfigTest extends GdxMockedTest {
     @Test
     public void testConfigWhenFileDoesNotExist() {
         AdvConfig config = new AdvConfig(TEST_WRITE_FILE);
+        Assert.assertEquals(ScreenResolution.RES_1024_768, config.getResolution());
+    }
+
+    @Test
+    public void testReadBadConfig() {
+        AdvConfig config = new AdvConfig(BAD_TEST_FILE);
         Assert.assertEquals(ScreenResolution.RES_1024_768, config.getResolution());
     }
